requirements.txt
Flask
python-dotenv
requests
pydantic
reportlab
python-dateutil
Pillow
gunicorn
waitress
typing_extensions
